terminal-notifier: fix installation
Instead of writing the bytes to disk and unpack them by reading
resulting file, pass a *zip.Reader to the responsible code.

This fixes a problem with previous code where gosx-notifier would not
have the permissions to write to current working directory (in some
invocation) and fail.

Since we don't write to disk anymore it's not an issue.